requirements.txt
Flask==3.0.0
gunicorn
Flask-SQLAlchemy
rq==1.15.0
redis
requests
python-dotenv
supervisor
PyYAML
google-generativeai
openai